+++ src/app/components/GoogleGeminiEffectDemo.tsx
